Split STSimulate.hs into smaller files in SimulatorLib directory. Still need documentation.
